Made map abstracted from graphics
Needed to do some fiddling to be able to test the mapcamera so it might
not be the best solution but again it should never reach production
since it is headless specific code
