+++ src/main/java/no/uib/inf112/core/map/MapHandler.java
@@ -5,11 +5,13 @@ import com.badlogic.gdx.graphics.OrthographicCamera;
 import com.badlogic.gdx.graphics.g2d.Batch;
 import com.badlogic.gdx.maps.MapProperties;
 import com.badlogic.gdx.maps.tiled.TiledMapTileSets;
+import com.badlogic.gdx.math.Vector2;
 import com.badlogic.gdx.utils.Disposable;
 import no.uib.inf112.core.player.Entity;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
+import java.util.Map;
 import java.util.Set;
 
 /**
@@ -103,7 +105,7 @@ public interface MapHandler extends Disposable {
      * @return A read-only set of known robots in the order they were added
      */
     @NotNull
-    Set<Entity> getEntities();
+    Map<Entity, Vector2> getEntities();
 
     /**
      * @param x The x coordinate to test
