+++ src/main/java/no/uib/inf112/core/map/TiledMapHandler.java
@@ -9,6 +9,7 @@ import com.badlogic.gdx.maps.tiled.TiledMapTileSets;
 import com.badlogic.gdx.maps.tiled.TmxMapLoader;
 import com.badlogic.gdx.maps.tiled.renderers.OrthogonalTiledMapRenderer;
 import com.badlogic.gdx.math.Vector2;
+import com.badlogic.gdx.utils.Disposable;
 import no.uib.inf112.core.player.Entity;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
@@ -18,65 +19,17 @@ import java.util.Map;
 import java.util.Set;
 
 
-public class TiledMapHandler extends MapCamera {
+public class TiledMapHandler extends MapCamera implements Disposable {
 
-    private TiledMap tiledMap;
     private OrthogonalTiledMapRenderer renderer;
-
-    private final TiledMapTileLayer boardLayer;
-    private final TiledMapTileLayer entityLayer;
-
-
-    //A map of all know entities and their last know location
-    private Map<Entity, Vector2> entities;
-
-    private int mapWidth;
-    private int mapHeight;
-    private int tileWidth;
-    private int tileHeight;
-
 
     /**
      * @param map The relative path from assets folder to the Tiled map file
      * @throws IllegalArgumentException if max zoom is less than min zoom
      */
     public TiledMapHandler(String map) {
-        super();
-
-        try {
-            TmxMapLoader.Parameters params = new TmxMapLoader.Parameters();
-            params.textureMagFilter = Texture.TextureFilter.Linear;
-            params.textureMinFilter = Texture.TextureFilter.Linear;
-            tiledMap = new TmxMapLoader().load(map, params);
-        } catch (final Exception e) {
-            throw new IllegalArgumentException("Failed to load map at '" + map + "'");
-        }
-
-        mapWidth = tiledMap.getProperties().get("width", int.class);
-        mapHeight = tiledMap.getProperties().get("height", int.class);
-        tileWidth = tiledMap.getProperties().get("tilewidth", int.class);
-        tileHeight = tiledMap.getProperties().get("tileheight", int.class);
-
-
-        TiledMapTileLayer baseLayer = null;
-        try {
-            baseLayer = (TiledMapTileLayer) tiledMap.getLayers().get(BOARD_LAYER_NAME);
-        } catch (ClassCastException ignore) {
-        }
-        if (baseLayer == null) {
-            throw new IllegalStateException("Given tiled map does not have a tile layer named '" + BOARD_LAYER_NAME + "'");
-        }
-        boardLayer = baseLayer;
-
-        //create a new empty layer for all the robots to play on :)
-        entityLayer = new TiledMapTileLayer(mapWidth, mapHeight, tileWidth, tileHeight);
-        tiledMap.getLayers().add(entityLayer);
-
-        renderer = new OrthogonalTiledMapRenderer(tiledMap);
-
-        //use a linked hashmap to make sure the iteration is consistent
-        entities = new LinkedHashMap<>();
-
+        super(map);
+        renderer = new OrthogonalTiledMapRenderer(getTiledMap());
     }
 
 
@@ -90,7 +43,7 @@ public class TiledMapHandler extends MapCamera {
 
     @Override
     public void update(float delta) {
-        for (Map.Entry<Entity, Vector2> entry : entities.entrySet()) {
+        for (Map.Entry<Entity, Vector2> entry : getEntities().entrySet()) {
 
             //make sure the new x and y are always consistent
             int x = entry.getKey().getX();
@@ -105,7 +58,7 @@ public class TiledMapHandler extends MapCamera {
                 continue;
             }
 
-            entityLayer.setCell((int) lastPos.x, (int) lastPos.y, null);
+            getEntityLayer().setCell((int) lastPos.x, (int) lastPos.y, null);
             setEntityOnBoard(entry.getKey(), lastPos, x, y);
 
         }
@@ -113,75 +66,9 @@ public class TiledMapHandler extends MapCamera {
 
 
     @Override
-    public MapProperties getProperties() {
-        return tiledMap.getProperties();
-    }
-
-    @NotNull
-    @Override
-    public TileType getBoardLayerTile(int x, int y) {
-        int tileId = boardLayer.getCell(x, y).getTile().getId();
-        return TileType.fromTiledId(tileId);
-    }
-
-    @Override
-    @Nullable
-    public Entity getEntity(int x, int y) {
-        Vector2 v = new Vector2(x, y);
-        for (Map.Entry<Entity, Vector2> entry : entities.entrySet()) {
-            if (v.equals(entry.getValue())) {
-                return entry.getKey();
-            }
-        }
-        return null;
-    }
-
-    @NotNull
-    @Override
-    public TiledMapTileSets getMapTileSets() {
-        return tiledMap.getTileSets();
-    }
-
-    @Override
-    public void addEntity(@NotNull Entity entity) {
-        for (Entity knownRobot : getEntities()) {
-            if (entity.getX() == knownRobot.getX() && entity.getY() == knownRobot.getY()) {
-                throw new IllegalStateException("Cannot add an entity on top of another entity");
-            }
-        }
-        entities.put(entity, null);
-    }
-
-    @Override
-    public boolean removeEntity(Entity entity) {
-        return entities.remove(entity) == null;
-    }
-
-    @NotNull
-    @Override
-    public Set<Entity> getEntities() {
-        return entities.keySet();
-    }
-
-    @Override
-    public boolean isOutsideBoard(int x, int y) {
-        return x < 0 || x >= getMapWidth() || y < 0 | y >= getMapHeight();
-    }
-
-    @Override
-    public int getMapWidth() {
-        return mapWidth;
-    }
-
-    @Override
-    public int getMapHeight() {
-        return mapHeight;
-    }
-
-    @Override
     public void dispose() {
         renderer.dispose();
-        tiledMap.dispose();
+        getTiledMap().dispose();
     }
 
 
@@ -201,7 +88,7 @@ public class TiledMapHandler extends MapCamera {
             throw new IllegalArgumentException("Given location (" + x + ", " + y + ") is out of bounds");
         }
         TiledMapTileLayer.Cell cell = new TiledMapTileLayer.Cell().setTile(entity.getTileType().getTile());
-        entityLayer.setCell(x, y, cell);
+        getEntityLayer().setCell(x, y, cell);
 
         oldPos.x = x;
         oldPos.y = y;
