+++ src/main/java/no/uib/inf112/desktop/Main.java
@@ -8,7 +8,7 @@ import no.uib.inf112.core.RoboRally;
 
 public class Main {
 
-    public static final boolean HEADLESS = false;
+    public static boolean HEADLESS = false;
 
     public static void main(String[] args) {
         LwjglApplicationConfiguration cfg = new LwjglApplicationConfiguration();
